more: remove function like preprocessor defines
This sort of abstraction makes code readability worse without much of gain.
Sure calling a function in these cases can be tiny bit slower, but who ever
claims that is forgetting more(1) is primarily an interactive pager and the
slowness originates to humans using it.  That in mind adding few
milliseconds of slowness here and there _does not matter_.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>