tailf: check printing criteria more carefully
Earlier size check did not print out contents of a file if it happen to
be same size as last time roll_file() function ran.  The earlier buggy
behavior was easy to reproduce with two terminal windows.  Contents of
the second echo were not visible.

term1> tailf /tmp/tailf-test
terminal2> echo 1 >| /tmp/tailf-test
terminal2> echo 2 >| /tmp/tailf-test

In this commit also a mount point and a file changes are detected,
leading to full printout of a file in such cases.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>