tailf: ensure file argument really is a file
The tailf(1) never worked very well with block or character devices,
sockets, fifos and such.  Now after mmap() is used to find last lines
even the little command used to work is broken, so test the tailf is
asked to follow a file and when not fail.  That said symlinks are OK, as
long they point to a file.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>